restructure team bindings query
