+++ pkg/registry/apis/identity/legacy/legacy_sql.go
@@ -25,135 +25,6 @@ type legacySQLStore struct {
 	sql legacysql.LegacyDatabaseProvider
 }
 
-// ListTeams implements LegacyIdentityStore.
-func (s *legacySQLStore) ListTeams(ctx context.Context, ns claims.NamespaceInfo, query ListTeamQuery) (*ListTeamResult, error) {
-	// for continue
-	query.Pagination.Limit += 1
-	query.OrgID = ns.OrgID
-	if ns.OrgID == 0 {
-		return nil, fmt.Errorf("expected non zero orgID")
-	}
-
-	sql, err := s.sql(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	req := newListTeams(sql, &query)
-	q, err := sqltemplate.Execute(sqlQueryTeamsTemplate, req)
-	if err != nil {
-		return nil, fmt.Errorf("execute template %q: %w", sqlQueryTeamsTemplate.Name(), err)
-	}
-
-	rows, err := sql.DB.GetSqlxSession().Query(ctx, q, req.GetArgs()...)
-	defer func() {
-		if rows != nil {
-			_ = rows.Close()
-		}
-	}()
-
-	res := &ListTeamResult{}
-	if err != nil {
-		return nil, err
-	}
-
-	var lastID int64
-	for rows.Next() {
-		t := team.Team{}
-		err = rows.Scan(&t.ID, &t.UID, &t.Name, &t.Email, &t.Created, &t.Updated)
-		if err != nil {
-			return res, err
-		}
-
-		lastID = t.ID
-		res.Teams = append(res.Teams, t)
-		if len(res.Teams) > int(query.Pagination.Limit)-1 {
-			res.Teams = res.Teams[0 : len(res.Teams)-1]
-			res.Continue = lastID
-			break
-		}
-	}
-
-	if query.UID == "" {
-		res.RV, err = sql.GetResourceVersion(ctx, "team", "updated")
-	}
-
-	return res, err
-}
-
-// ListTeamsBindings implements LegacyIdentityStore.
-func (s *legacySQLStore) ListTeamBindings(ctx context.Context, ns claims.NamespaceInfo, query ListTeamBindingsQuery) (*ListTeamBindingsResult, error) {
-	// for continue
-	query.Pagination.Limit += 1
-	query.OrgID = ns.OrgID
-	if query.OrgID == 0 {
-		return nil, fmt.Errorf("expected non zero orgID")
-	}
-
-	sql, err := s.sql(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	req := newListTeamBindings(sql, &query)
-	q, err := sqltemplate.Execute(sqlQueryTeamBindings, req)
-	if err != nil {
-		return nil, fmt.Errorf("execute template %q: %w", sqlQueryTeamsTemplate.Name(), err)
-	}
-
-	rows, err := sql.DB.GetSqlxSession().Query(ctx, q, req.GetArgs()...)
-	defer func() {
-		if rows != nil {
-			_ = rows.Close()
-		}
-	}()
-
-	if err != nil {
-		return nil, err
-	}
-
-	res := &ListTeamBindingsResult{}
-	grouped := map[string][]TeamMember{}
-
-	var lastID int64
-	var atTeamLimit bool
-
-	for rows.Next() {
-		m := TeamMember{}
-		err = rows.Scan(&m.ID, &m.TeamUID, &m.TeamID, &m.UserUID, &m.Created, &m.Updated, &m.Permission)
-		if err != nil {
-			return res, err
-		}
-
-		lastID = m.TeamID
-		members, ok := grouped[m.TeamUID]
-		if ok {
-			grouped[m.TeamUID] = append(members, m)
-		} else if !atTeamLimit {
-			grouped[m.TeamUID] = []TeamMember{m}
-		}
-
-		if len(grouped) >= int(query.Pagination.Limit)-1 {
-			atTeamLimit = true
-			res.Continue = lastID
-		}
-	}
-
-	if query.UID == "" {
-		res.RV, err = sql.GetResourceVersion(ctx, "team_member", "updated")
-	}
-
-	res.Bindings = make([]TeamBinding, 0, len(grouped))
-	for uid, members := range grouped {
-		res.Bindings = append(res.Bindings, TeamBinding{
-			TeamUID: uid,
-			Members: members,
-		})
-	}
-
-	return res, err
-}
-
 // ListTeamMembers implements LegacyIdentityStore.
 func (s *legacySQLStore) ListTeamMembers(ctx context.Context, ns claims.NamespaceInfo, query ListTeamMembersQuery) (*ListTeamMembersResult, error) {
 	query.Pagination.Limit += 1
